Loadable Plan - Excel Export - Functionality comments (#1347)
* Committing for Jira #DSS-3026

* Committing for Jira #2976

* Committing to fix functionality comments for Loadable Plan Excel Export

* Committing for Jira #2170

* Committing for Functionality comment fix loadable plan excel export

+++ api/cp-dss-api/loadable-study/src/main/java/com/cpdss/loadablestudy/domain/VesselTanksTable.java
@@ -9,11 +9,11 @@ import lombok.Data;
 public class VesselTanksTable {
   private String tankNo;
   private String cargoCode;
-  private float ullage;
-  private float loadedPercentage;
-  private float shipsNBbls;
-  private float shipsMt;
-  private float shipsKlAt15C;
+  private double ullage;
+  private double loadedPercentage;
+  private double shipsNBbls;
+  private double shipsMt;
+  private double shipsKlAt15C;
   private float frameNoFrom;
   private float frameNoTo;
   private String colorCode;
